requirement.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-cities==0.6.2
django-model-utils==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.0
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
swapper==1.4.0
tqdm==4.67.3
urllib3==2.6.3
